common functionality of ImageUploadType and FileUploadType has been moved to AbstractFileUploadType
