+++ project-base/config/packages/twig.yaml
@@ -4,7 +4,7 @@ twig:
     form_themes:
         - '@ShopsysFramework/Common/Form/theme.html.twig'
         - '@ShopsysFramework/Admin/Form/colorpickerFields.html.twig'
-        - '@ShopsysFramework/Admin/Form/fileuploadFields.html.twig'
+        - '@ShopsysFramework/Admin/Form/abstractFileuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/imageuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/pricetableFields.html.twig'
         - '@ShopsysFramework/Admin/Form/orderItems.html.twig'
